zw - wip, moving on to anotther approach
Currently thinking it is better to create separate in_controller files
for praxis and rails and not try to maintain an abstract interface.
This allows for changes related to updates to rails and updates to
praxis to be accounted for indepenently.  It seems like declarative
authorization itself hasn't changed much in year.