Add a rubocop for filter_access_to placement
There are a number of controllers in property app that
put the filter_access_to statement before some of the
before_actions. This is explicitly not recommended in the
documentation for the declarative auth gem, but is not
enforced.

Apps that import this cop will have its controller files
monitored for this violation by default.