Update rubocop rule for accuracy
The previous design of the rule was flawed, in that
not *every* before_action needs to precede the filter
statement. And the concern is only if the filter
statement has attribute_check enabled and no load method.

In such a case, the object being evaluated is implicit
and the filter may still be valid. So the rule will
mark the filter, but just as a warning so the dev is
aware and might choose to make it explicit.